third_party/blink/web_tests/external/wpt/css/css-animations/parsing/keyframes-name-invalid-expected.txt

This is a testharness.js-based test.
[FAIL] invalid: @keyframes none { }
  assert_equals: expected 0 but got 1
[FAIL] invalid: @keyframes default { }
  assert_equals: expected 0 but got 1
[FAIL] invalid: @keyframes initial { }
  assert_equals: expected 0 but got 1
[FAIL] invalid: @keyframes inherit { }
  assert_equals: expected 0 but got 1
[FAIL] invalid: @keyframes unset { }
  assert_equals: expected 0 but got 1
[FAIL] invalid: @keyframes revert { }
  assert_equals: expected 0 but got 1
[FAIL] invalid: @keyframes revert-layer { }
  assert_equals: expected 0 but got 1
[PASS] invalid: @keyframes 12 { }
[PASS] invalid: @keyframes -12 { }
[PASS] invalid: @keyframes 12foo { }
[PASS] invalid: @keyframes foo.bar { }
[PASS] invalid: @keyframes one two { }
[PASS] invalid: @keyframes one, two { }
[PASS] invalid: @keyframes one, initial { }
[PASS] invalid: @keyframes one, inherit { }
[PASS] invalid: @keyframes one, unset { }
[PASS] invalid: @keyframes default, two { }
[PASS] invalid: @keyframes revert, three { }
[PASS] invalid: @keyframes revert-layer, four { }
Harness: the test ran to completion.
